feat: implement private repo friendly cloning logic
Copying the repository in the container and then checkout is safer for
case where the repository is private. It circumvents the
problematic of leaked secret that can persist in build history and
cached layer. Since the repository is already cloned on the host
machine whether it's a personal computer, a robot or a CI server its
safer and easier to manage to delegate the responsibility of secret
management to the user.

- Fix Dockerfile template by simplifying directory check logic.
- Implement additional dev workspace-related logic.
- Update docker-compose build config for local source path.
- Improve error messaging in helper scripts.
- Set default dev branch for DN project deploy in env file.
- Add README for utilities with Git submodule usage notes and references.

Issue NMO-558